Add functionality for company deletion from the system
A new method was added in CompanyList.java that allows specific companies to be removed using their filename. These changes also included modifications to the UI in Dashboard.java so that users have the option to delete a company. Update in Company.java to keep track of the fileName also aids the deletion operation. Moreover, the Authentication.java file was refactored for logging in and sign up process, primarily for checkUserExists method with better IO Error Messages.