Add logging and handle exceptions in Dashboard and Company classes
This commit introduces logging to the Dashboard and Company classes and improves exception handling. The changes include the incorporation of a log message when an UnsupportedLookAndFeelException is triggered in the Dashboard class, and the handling of an ArrayIndexOutOfBoundsException in the Company class. It also adjusts the JDK version in misc.xml and simplifies the process of checking if data is complete in the Statistic class.