Add setters in Data.java, refactor Statistic, Company classes
This commit adds setter methods in the Data model, and refactors Statistic and Company classes. In the Statistic class, the extrapolation method is modified to extrapolate for years instead of months. The Company class modifications include adjusting the logic for extrapolating and summing data. Also, some fixes and improvements are made in the Merger class.